requirements_with_versions.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
fastapi==0.115.6
googleapis-common-protos==1.66.0
grpcio==1.68.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
MarkupSafe==3.0.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
protobuf==5.29.2
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.0
zipp==3.21.0
